Split the padding-invariance tests by source module
From the CodeRabbit review: one test file per source file. The encoder
checks move into the existing conformer, branchformer, e_branchformer and
multiconvformer test files, sharing a small helper module; the convolution
module and the cgMLP unit get their own files. The cgMLP test gives the
convolution real weights, since its near-identity initialisation hid the
leak it tests for.
